Renamed to java-aws-lambda-pojos
