+++ java-aws-lambda-pojos/src/main/scala/models/generator/javaAwsLambdaPojos/ApidocComments.scala
@@ -1,4 +1,4 @@
-package models.generator.javapojos
+package models.generator.javaAwsLambdaPojos
 
 import lib.Constants
