+++ java-aws-lambda-pojos/src/main/scala/models/generator/javaAwsLambdaPojos/JavaAwsLambdaPOJOClasses.scala
@@ -1,7 +1,7 @@
-package models.generator.javapojos
+package models.generator.javaAwsLambdaPojos
 
-object JavaPOJOClasses
-  extends BaseJavaPOJOCodeGenerator {
+object JavaAwsLambdaPOJOClasses
+  extends BaseJavaAwsLambdaPOJOCodeGenerator {
   override def getJavaDocFileHeader() = "WARNING: not all features (notably unions) and data types work with the java generator yet.  \n" +
     "Java POJO generator is designed to be used with AWS Lambdas.  \n" +
     "If you are considering using this library, would like to request/discuss features, or would like to share how you're using it, please contact daniel.kirby@gilt.com \n"
